chore: modify structure State View
