+++ ui/src/composables/use-robot-state.ts
@@ -6,6 +6,6 @@ export function useQueryRobotState(opt?: Partial<AxiosRequestConfig>, refetchInt
   return useQuery({
     queryKey: ['robotState'],
     queryFn: () => robotStateAPI.getRobotState(opt),
-    refetchInterval: refetchInterval || false,
+    refetchInterval,
   })
 }
